more changes to support default cubes looping
