+++ jwst/cube_build/cube_build.py
@@ -10,21 +10,11 @@ import logging
 
 from astropy.io import fits
 from astropy.modeling import models
-#from gwcs.utils import _domain_to_bounds
 from ..associations import Association
 from .. import datamodels
-from ..assign_wcs import nirspec
-from ..assign_wcs import pointing
 from . import cube_build_io_util
-from . import cube_build_wcs_util
 from . import file_table
 from . import instrument_defaults
-from . import spaxel
-from . import cube_overlap
-from . import cube_cloud
-from . import data_types
-
-from gwcs import wcstools
 
 
 log = logging.getLogger(__name__)
@@ -34,10 +24,10 @@ class CubeData(object):
 # CubeData - holds all the importatn informtion for IFU Cube Building:
 # wcs, data, reference data
 
-    def __init__(self, cube_type,
+    def __init__(self, 
+                 cube_type,
                  input_models,
                  input_filenames,
-                 output_name_base,
                  data_type,
                  par_filename,
                  resol_filename,
@@ -46,7 +36,6 @@ class CubeData(object):
         self.cube_type = cube_type
         self.input_models = input_models
         self.input_filenames = input_filenames
-        self.output_name_base = output_name_base
         self.data_type = data_type
         self.par_filename = par_filename
         self.resol_filename = resol_filename
@@ -56,36 +45,14 @@ class CubeData(object):
         self.subchannel = pars.get('subchannel')
         self.grating = pars.get('grating')
         self.filter = pars.get('filter')
-        self.scale1 = pars.get('scale1')
-        self.scale2 = pars.get('scale2')
-        self.scalew = pars.get('scalew')
-        self.rois = pars.get('rois')
-        self.roiw = pars.get('roiw')
-        self.output_file = pars.get('output_file')
-        self.interpolation = pars.get('interpolation')
-        self.coord_system = pars.get('coord_system')
         self.offset_list = pars.get('offset_list')
-        self.wavemin = pars.get('wavemin')
-        self.wavemax = pars.get('wavemax')
         self.weighting = pars.get('weighting')
-        self.weight_power = pars.get('weight_power')
-        self.xdebug = pars.get('xdebug')
-        self.ydebug = pars.get('ydebug')
-        self.zdebug = pars.get('zdebug')
-        self.debug_pixel = pars.get('debug_pixel')
-        self.spaxel_debug = pars.get('spaxel_debug')
         self.output_type = pars.get('output_type')
 
         self.ra_offset = []  # units arc seconds
         self.dec_offset = [] # units arc seconds
         self.detector = None
         self.instrument = None
-        self.num_bands = 0
-
-        self.band_channel = []
-        self.band_subchannel = []
-        self.band_filter = []
-        self.band_grating = []
 
         self.all_channel = []
         self.all_subchannel =[]
@@ -93,34 +60,6 @@ class CubeData(object):
         self.all_filter = []
 
         self.output_name = ''
-        self.number_files = 0
-
-        self.Cdelt1 = None
-        self.Cdelt2 = None
-        self.Cdelt3 = None
-        self.Crpix1 = None
-        self.Crpix2 = None
-        self.Crpix3 = None
-        self.Crval1 = None
-        self.Crval2 = None
-        self.Crval3 = None
-        self.naxis1 = None
-        self.naxis2 = None
-        self.naxis3 = None
-
-        self.a_min = 0
-        self.a_max = 0
-        self.b_min = 0
-        self.b_max = 0
-        self.lambda_min = 0
-        self.lambda_max = 0
-        self.xcoord = None
-        self.ycoord = None
-        self.zcoord = None
-
-        self.spaxel = []        # list of spaxel classes
-#********************************************************************************
-
 #********************************************************************************
     def setup(self):
 
@@ -173,7 +112,6 @@ class CubeData(object):
         num = 0
         num = len(self.input_filenames)
         
-        self.number_files = num
         self.detector = detector
         self.instrument = instrument
 #________________________________________________________________________________
@@ -181,28 +119,7 @@ class CubeData(object):
     # constructed from.
     # fills in band_channel, band_subchannel, band_grating, band_filer
 #________________________________________________________________________________
-        cube_build_io_util.determine_band_coverage(self, master_table)
-
-        # temp code
-        self.band_channel = self.all_channel
-        self.band_subchannel = self.all_subchannel
-        self.band_grating = self.all_grating
-        self.band_filter = self.all_filter
-        print('*****band channel*****',self.band_channel)
-#________________________________________________________________________________
-
-    # check on interpolation = area and coord_system=alpha-beta types of cubes
-    # if interpolation = area also checks that the use did not supply a scale2
-    # values (beta dim)
-#________________________________________________________________________________
-### move these routines  to individual cube checking 
-
-###        cube_build_io_util.check_cube_type(self)
-###
-###        self.output_name = cube_build_io_util.update_output_name(self)
-###        if not self.single:
-###            log.info('Output Name %s',self.output_name)
-
+        CubeData.determine_band_coverage(self, master_table)
 #________________________________________________________________________________
 # InstrumentDefaults is an  dictionary that holds default parameters for
 # difference instruments and for each band
@@ -222,31 +139,152 @@ class CubeData(object):
             log.info('Reading default MIRI cube resolution file %s', self.resol_filename)
             cube_build_io_util.read_resolution_file(self,instrument_info)
 #________________________________________________________________________________
-# get the ROI sizes
         self.instrument_info = instrument_info
-        roi = CubeData.determine_roi_size(self)
-        # if the user has not set the size of the ROI then use defaults in reference
-        # parameter file
-
-        if self.roiw == 0.0: self.roiw = roi[0]
-        if self.rois == 0.0:  # user did not set so use defaults 
-            self.rois = roi[1]
-            if self.single or self.number_files < 4:
-                self.rois = self.rois * 1.5
-                log.info('Increasing spatial region of interest' + \
-                             ' default value set for 4 dithers %f', \
-                             self.rois)
-        if self.interpolation == 'pointcloud':
-            log.info('Region of interest  %f %f',self.rois,self.roiw)
-
 #________________________________________________________________________________
 # Set up values to return and acess for other parts of cube_build
 
         self.master_table = master_table
-        
-        return self.output_file
+
+        return {'instrument':self.instrument,'detector':self.detector,
+                'instrument_info':self.instrument_info,
+                'master_table':self.master_table}
+
 
 #********************************************************************************
+    def determine_band_coverage(self, master_table):
+#********************************************************************************
+        """
+        Short Summary
+        -------------
+        Function to determine which files contain channels and subchannels are used
+        in the creation of the cubes.
+        For MIRI The channels  to be used are set by the association and the
+        subchannels are  determined from the data
+
+        Parameter
+        ----------
+        self containing user set input parameters:
+        self.channel, self.subchannel
+
+        Returns
+        -------
+        fills in self.band_channel, self.band_subchannel
+        self.band_grating, self.band_filter
+
+        """
+#________________________________________________________________________________
+# IF INSTRUMENT = MIRI
+# loop over the file names
+
+        if self.instrument == 'MIRI':
+            ValidChannel = ['1', '2', '3', '4']
+            ValidSubchannel = ['SHORT', 'MEDIUM', 'LONG']
+
+            nchannels = len(ValidChannel)
+            nsubchannels = len(ValidSubchannel)
+#________________________________________________________________________________
+        # for MIRI we can set the channel and subchannel
+            user_clen = len(self.channel)
+            user_slen = len(self.subchannel)
+#________________________________________________________________________________
+            for i in range(nchannels):
+                for j in range(nsubchannels):
+                    nfiles = len(master_table.FileMap['MIRI'][ValidChannel[i]][ValidSubchannel[j]])
+                    if nfiles > 0 :
+#________________________________________________________________________________
+        # neither parameters not set
+                        if user_clen == 0 and  user_slen == 0:
+                            self.all_channel.append(ValidChannel[i])
+                            self.all_subchannel.append(ValidSubchannel[j])
+#________________________________________________________________________________
+# channel was set by user but not sub-channel
+                        elif user_clen !=0 and user_slen ==0:
+                        # now check if this channel was set by user
+                            if (ValidChannel[i] in self.channel ):
+                                self.all_channel.append(ValidChannel[i])
+                                self.all_subchannel.append(ValidSubchannel[j])
+#________________________________________________________________________________
+# sub-channel was set by user but not channel
+                        elif user_clen ==0 and user_slen !=0:
+                            if (ValidSubchannel[j] in self.subchannel):
+                                self.all_channel.append(ValidChannel[i])
+                                self.all_subchannel.append(ValidSubchannel[j])
+#________________________________________________________________________________
+# both parameters set
+                        else:
+                            if (ValidChannel[i] in self.channel and
+                                ValidSubchannel[j] in self.subchannel):
+
+                                self.all_channel.append(ValidChannel[i])
+                                self.all_subchannel.append(ValidSubchannel[j])
+
+            log.info('The desired cubes covers the MIRI Channels: %s',
+                     self.all_channel)
+            log.info('The desired cubes covers the MIRI subchannels: %s',
+                     self.all_subchannel)
+
+            number_channels = len(self.all_channel)
+            number_subchannels = len(self.all_subchannel)
+
+            if number_channels == 0:
+                raise ErrorNoChannels(
+                    "The cube  does not cover any channels, change parameter channel")
+            if number_subchannels == 0:
+                raise ErrorNoSubchannels(
+                    "The cube  does not cover any subchannels, change parameter subchannel")
+
+#______________________________________________________________________
+        if self.instrument == 'NIRSPEC':
+        # 1 to 1 mapping VALIDGWA[i] -> VALIDFWA[i]
+            ValidGWA = ['G140M', 'G140H', 'G140M', 'G140H', 'G235M', 'G235H',
+                        'G395M', 'G395H', 'PRISM']
+            ValidFWA = ['F070LP', 'F070LP', 'F100LP', 'F100LP', 'F170LP',
+                        'F170LP', 'F290LP', 'F290LP', 'CLEAR']
+
+            nbands = len(ValidFWA)
+#________________________________________________________________________________
+        # check if input filter or grating has been set
+            user_glen = len(self.grating)
+            user_flen = len(self.filter)
+
+            if user_glen ==0 and user_flen !=0:
+                raise ErrorMissingParameter("Filter specified, but Grating was not")
+
+            if user_glen !=0 and user_flen ==0:
+                raise ErrorMissingParameter("Grating specified, but Filter was not")
+        # Grating and Filter not set - read in from files and create a list of all
+        # the filters and grating contained in the files
+            if user_glen == 0 and  user_flen == 0:
+                for i in range(nbands):
+
+                    nfiles = len(master_table.FileMap['NIRSPEC'][ValidGWA[i]][ValidFWA[i]])
+                    if nfiles > 0:
+                        self.all_grating.append(ValidGWA[i])
+                        self.all_filter.append(ValidFWA[i])
+
+        # Both filter and grating input parameter have been set
+        # Find the files that have these parameters set
+
+            else:
+                for i in range(nbands):
+                    nfiles = len(master_table.FileMap['NIRSPEC'][ValidGWA[i]][ValidFWA[i]])
+                    if nfiles > 0:
+                        # now check if THESE Filter and Grating input parameters were set
+                        if (ValidFWA[i] in self.filter and
+                            ValidGWA[i] in self.grating):
+                            self.all_grating.append(ValidGWA[i])
+                            self.all_filter.append(ValidFWA[i])
+
+            number_filters = len(self.all_filter)
+            number_gratings = len(self.all_grating)
+
+            if number_filters == 0:
+                raise ErrorNoFilters("The cube  does not cover any filters")
+            if number_gratings == 0:
+                raise ErrorNoGratings("The cube  does not cover any gratings")
+
+
+#______________________________________________________________________                        
 # Determine the number of cubes going to create based on:
 # self.output_type
 # MIRI: self.band_channel, self.band_subchannel
@@ -262,8 +300,12 @@ class CubeData(object):
 # MIRI
 #______________________________________________________________________
         if self.instrument == 'MIRI':
-            band_channel = list(set(self.all_channel))
-            band_subchannel =list( set(self.all_subchannel))
+
+            band_channel = self.all_channel
+            band_subchannel =self.all_subchannel
+
+            print('all_channel ',band_channel)
+            print('all_subchannel',band_subchannel)
 #user and single
             if (self.output_type == 'user' or self.output_type =='single' or
                 self.output_type =='multi') :
@@ -287,31 +329,33 @@ class CubeData(object):
             if self.output_type == 'band':
                 log.info('Output Cubes are single channel, single sub-channel IFU Cubes')
                 for i in range(len(band_channel)):
-                    for j in range(len(band_subchannel)):
-                        num_cubes = num_cubes+1
-                        cube_no = str(num_cubes)
-                        cube_pars[cube_no] = {}
-                        cube_pars[cube_no]['pars1'] = {}
-                        cube_pars[cube_no]['pars2'] = {}
-                        this_channel  = []
-                        this_subchannel = []
-                        this_channel.append(band_channel[i])
-                        this_subchannel.append(band_subchannel[j])
-                        cube_pars[cube_no]['par1'] = this_channel
-                        cube_pars[cube_no]['par2'] = this_subchannel
+                    num_cubes = num_cubes+1
+                    cube_no = str(num_cubes)
+                    cube_pars[cube_no] = {}
+                    cube_pars[cube_no]['pars1'] = {}
+                    cube_pars[cube_no]['pars2'] = {}
+                    this_channel  = []
+                    this_subchannel = []
+                    this_channel.append(band_channel[i])
+                    this_subchannel.append(band_subchannel[i])
+                    cube_pars[cube_no]['par1'] = this_channel
+                    cube_pars[cube_no]['par2'] = this_subchannel
 
 # default channel cubes 
             if self.output_type  == 'channel':
                 log.info('Output cubes are single channel and all subchannels in data')
                 num_cubes = 0 
-                for i in range(len(band_channel)):
+                channel_no_repeat = list(set(band_channel))
+                for i in range(len(channel_no_repeat)):
                         num_cubes = num_cubes+1
                         cube_no = str(num_cubes)
                         cube_pars[cube_no] = {}
                         cube_pars[cube_no]['pars1'] = {}
                         cube_pars[cube_no]['pars2'] = {}
                         this_channel  = []
-                        this_subchannel = band_subchannel
+                        for j in range(band_channel):
+                            if j == i:
+                                this_subchannel = band_subchannel[j]
                         this_channel.append(i)
                         cube_pars[cube_no]['par1'] = this_channel
                         cube_pars[cube_no]['par2'] = this_subchannel
@@ -375,835 +419,9 @@ class CubeData(object):
                         
         list1 = cube_pars['1']['par1']
         list2 = cube_pars['1']['par2']
-        print('channel for cube 1',list1)
-        print('subchannel for cube 1',list2)
-
         self.num_cubes = num_cubes
         self.cube_pars = cube_pars
-
-        print(cube_pars['1']['par1'])
-        print(cube_pars['1']['par2'])
-
-#        print(cube_pars['2']['par1'])
-#        print(cube_pars['2']['par2'])
         return self.num_cubes,self.cube_pars
 #********************************************************************************
 
-    def setup_wcs(self):
 
-#********************************************************************************
-        """
-        Short Summary
-        -------------
-        Function to determine the min and max coordinates of the spectral
-        cube,given channel & subchannel
-
-
-        Parameter
-        ----------
-        self.master_table:  A table that contains the channel/subchannel or
-        filter/grating for each input file
-        self.instrument_info: Default information on the MIRI and NIRSPEC instruments.
-
-        Returns
-        -------
-        Cube Dimension Information:
-        Footprint of cube: min and max of coordinates of cube.
-        If an offset list is provided then these values are applied.
-        If the coordinate system is alpha-beta (MIRI) then min and max
-        coordinates of alpha (arc sec), beta (arc sec) and lambda (microns)
-        If the coordinate system is ra-dec then the min and max of
-        ra(degress), dec (degrees) and lambda (microns) is returned.
-        """
-
-#________________________________________________________________________________
-        if self.cube_type == 'File' or self.cube_type == 'ASN' :
-            log.info('Building Cube %s ', self.output_name)
-
-        # Scale is 3 dimensions and is determined from values held in  instrument_info.GetScale
-        scale = cube_build_wcs_util.determine_scale(self)
-        self.Cdelt1 = scale[0]
-        self.Cdelt2 = scale[1]
-        self.Cdelt3 = scale[2]
-
-        if self.instrument == 'MIRI':
-            parameter1 = self.band_channel
-            parameter2 = self.band_subchannel
-        elif self.instrument == 'NIRSPEC':
-            parameter1 = self.band_grating
-            parameter2 = self.band_filter
-
-        a_min = []
-        a_max = []
-        b_min = []
-        b_max = []
-        lambda_min = []
-        lambda_max = []
-
-        log.info('Number of bands in cube  %i',self.num_bands)
-        print('band channel',self.band_channel)
-        for i in range(self.num_bands):
-            this_a = parameter1[i]
-            this_b = parameter2[i]
-            log.debug('Working on data  from %s,%s',this_a,this_b)
-            n = len(self.master_table.FileMap[self.instrument][this_a][this_b])
-            log.debug('number of files %d ', n)
-    # each file find the min and max a and lambda (OFFSETS NEED TO BE APPLIED TO THESE VALUES)
-            for k in range(n):
-                amin = 0.0
-                amax = 0.0
-                bmin = 0.0
-                bmax = 0.0
-                lmin = 0.0
-                lmax = 0.0
-                c1_offset = 0.0
-                c2_offset = 0.0
-                ifile = self.master_table.FileMap[self.instrument][this_a][this_b][k]
-                ioffset = len(self.master_table.FileOffset[this_a][this_b]['C1'])
-                if ioffset == n:
-                    c1_offset = self.master_table.FileOffset[this_a][this_b]['C1'][k]
-                    c2_offset = self.master_table.FileOffset[this_a][this_b]['C2'][k]
-#________________________________________________________________________________
-# Open the input data model
-# Find the footprint of the image
-
-                with datamodels.ImageModel(ifile) as input_model:
-                    if self.instrument == 'NIRSPEC':
-                        flag_data = 0
-                        ch_footprint = cube_build_wcs_util.find_footprint_NIRSPEC(self,
-                                                                              input_model,
-                                                                              flag_data)
-                        amin, amax, bmin, bmax, lmin, lmax = ch_footprint
-#________________________________________________________________________________
-                    if self.instrument == 'MIRI':
-                        ch_footprint = cube_build_wcs_util.find_footprint_MIRI(self,
-                                                                           input_model,
-                                                                           this_a,
-                                                                           self.instrument_info)
-                        amin, amax, bmin, bmax, lmin, lmax = ch_footprint
-
-# If a dither offset list exists then apply the dither offsets (offsets in arc seconds)
-
-                    amin = amin - c1_offset/3600.0
-                    amax = amax - c1_offset/3600.0
-
-                    bmin = bmin - c2_offset/3600.0
-                    bmax = bmax - c2_offset/3600.0
-
-                    a_min.append(amin)
-                    a_max.append(amax)
-                    b_min.append(bmin)
-                    b_max.append(bmax)
-                    lambda_min.append(lmin)
-                    lambda_max.append(lmax)
-#________________________________________________________________________________
-    # done looping over files determine final size of cube
-
-        final_a_min = min(a_min)
-        final_a_max = max(a_max)
-        final_b_min = min(b_min)
-        final_b_max = max(b_max)
-        final_lambda_min = min(lambda_min)
-        final_lambda_max = max(lambda_max)
-
-        if(self.wavemin != None and self.wavemin > final_lambda_min):
-            final_lambda_min = self.wavemin
-            log.info('Changed min wavelength of cube to %f ',final_lambda_min)
-
-        if(self.wavemax != None and self.wavemax < final_lambda_max):
-            final_lambda_max = self.wavemax
-            log.info('Changed max wavelength of cube to %f ',final_lambda_max)
-#________________________________________________________________________________
-        if self.instrument =='MIRI' and self.coord_system=='alpha-beta':
-        #  we have a 1 to 1 mapping in beta dimension.
-            nslice = self.instrument_info.GetNSlice(parameter1[0])
-            log.info('Beta Scale %f ',self.Cdelt2)
-            self.Cdelt2 = (final_b_max - final_b_min)/nslice
-        #print('remove********')
-        #Cube.Cdelt2 = 0.17722104
-            final_b_max = final_b_min + (nslice)*self.Cdelt2
-        #print('remove********')
-            log.info('Changed the Beta Scale dimension so we have 1 -1 mapping between beta and slice #')
-            log.info('New Beta Scale %f ',self.Cdelt2)
-#________________________________________________________________________________
-# Test that we have data (NIRSPEC NRS2 only has IFU data for 3 configurations)
-        test_a = final_a_max - final_a_min
-        test_b = final_b_max - final_b_min
-        test_w = final_lambda_max - final_lambda_min
-        tolerance1 = 0.00001
-        tolerance2 = 0.1
-        if(test_a < tolerance1 or test_b < tolerance1 or test_w < tolerance2):
-            log.info('No Valid IFU slice data found %f %f %f ',test_a,test_b,test_w)
-#________________________________________________________________________________
-        cube_footprint = (final_a_min, final_a_max, final_b_min, final_b_max,
-                     final_lambda_min, final_lambda_max)
-#________________________________________________________________________________
-    # Based on Scaling and Min and Max values determine naxis1, naxis2, naxis3
-    # set cube CRVALs, CRPIXs and xyz coords (center  x,y,z vector spaxel centers)
-
-        if(self.coord_system == 'ra-dec'):
-            cube_build_wcs_util.set_geometry(self,cube_footprint)
-        else:
-            cube_build_wcs_util.set_geometryAB(self,cube_footprint) # local coordinate system
-
-        cube_build_wcs_util.print_cube_geometry(self)
-
-
-#********************************************************************************
-
-    def build_ifucube(self):
-
-        """
-        Short Summary
-        -------------
-        Loop over every band contained in the IFU cube and read in the data associated with the band
-        Map the detector data to the cube output coordinate system
-
-        Parameter
-        ----------
-        spaxel - a list of spaxel members holding the detector flux information
-
-        Returns
-        -------
-        each spaxel element with the mapped detector values associated with it
-
-        """
-
-
-        CubeData.find_output_type(self)
-
-        self.spaxel = CubeData.create_spaxel(self)
-
-#        CubeData.SubtractBackground(self)
-
-        # now need to loop over every file that covers this channel/subchannel (MIRI)
-        # or Grating/filter(NIRSPEC)
-        #and map the detector pixels to the cube spaxel.
-        if(self.instrument == 'MIRI'):
-            parameter1 = self.band_channel   
-            parameter2 = self.band_subchannel
-        elif(self.instrument == 'NIRSPEC'):
-            parameter1 = self.band_grating
-            parameter2 = self.band_filter
-
-        number_bands = len(parameter1)
-        t0 = time.time()
-        for i in range(number_bands):
-            this_par1 = parameter1[i]
-            this_par2 = parameter2[i]
-
-            log.debug("Working on Band defined by:%s %s " ,this_par1,this_par2)
-            CubeData.map_detector_to_spaxel(self,this_par1, this_par2,self.spaxel)
-
-
-        t1 = time.time()
-        log.info("Time Map All slices on Detector to Cube = %.1f.s" % (t1 - t0,))
-#_______________________________________________________________________
-# Mapped all data to cube or Point Cloud
-# now determine Cube Spaxel flux
-
-        t0 = time.time()
-        CubeData.find_spaxel_flux(self, self.spaxel)
-
-        t1 = time.time()
-        log.info("Time find Cube Flux= %.1f.s" % (t1 - t0,))
-
-        IFUCube = CubeData.setup_IFUCube(self,0)
-
-#_______________________________________________________________________
-# shove Flux and iflux in the  final IFU cube
-        CubeData.update_IFUCube(self,IFUCube, self.spaxel)
-
-        return IFUCube
-
-#********************************************************************************
-
-    def build_ifucube_single(self):
-
-        """
-        Short Summary
-        -------------
-        Loop over every band contained in the IFU cube and read in the data associated with the band
-        Map the detector data to the cube output coordinate system
-
-        Parameter
-        ----------
-        spaxel - a list of spaxel members holding the detector flux information
-
-        Returns
-        -------
-        each spaxel element with the mapped detector values associated with it
-
-        """
-
-        # loop over input models
-
-        single_IFUCube = datamodels.ModelContainer()
-        n = len(self.input_models)
-        this_par1 = self.band_channel[0] # only one channel is used in this approach
-        this_par2 = None # not import for this type of mapping
-
-        self.weighting =='msm'
-        c1_offset = 0
-        c2_offset = 0
-        for j in range(n):
-            t0 = time.time()
-# for each new data model create a new spaxel
-            spaxel = []
-            spaxel = CubeData.create_spaxel(self)
-
-            with datamodels.ImageModel(self.input_models[j]) as input_model:
-#********************************************************************************
-# pulled necessary routines from   CubeData.map_detector_to_spaxel
-                if self.instrument == 'MIRI':
-#________________________________________________________________________________
-                    xstart, xend = self.instrument_info.GetMIRISliceEndPts(this_par1)
-                    y, x = np.mgrid[:1024, xstart:xend]
-                    y = np.reshape(y, y.size)
-                    x = np.reshape(x, x.size)
-
-                    cube_cloud.match_det2cube(self,input_model,
-                                              x, y, j,
-                                              this_par1,this_par2,
-                                              spaxel,
-                                              c1_offset, c2_offset)
-
-                elif instrument == 'NIRSPEC':
-                    # each file, detector has 30 slices - wcs information access seperately for each slice
-                    start_slice = 0
-                    end_slice = 29
-                    nslices = end_slice - start_slice + 1
-                    regions = list(range(start_slice, end_slice + 1))
-                    for ii in regions:
-                        t0a = time.time()
-                        slice_wcs = nirspec.nrs_wcs_set_input(input_model, ii)
-                        yrange = slice_wcs.bounding_box[1][0],slice_wcs.bounding_box[1][1]
-                        xrange = slice_wcs.bounding_box[0][0],slice_wcs.bounding_box[0][1]
-                        x,y = wcstools.grid_from_bounding_box(slice_wcs.bounding_box)
-
-
-                        cube_cloud.match_det2cube(self,input_model,
-                                                  x, y, ii,
-                                                  this_par1,this_par2,
-                                                  spaxel,
-                                                  c1_offset, c2_offset)
-
-                        t1a = time.time()
-                        log.debug("Time Match one NIRSPEC slice  to IFUCube = %.1f.s" % (t1a - t0a,))
-#_______________________________________________________________________
-# shove Flux and iflux in the  final IFU cube
-            CubeData.find_spaxel_flux(self, spaxel)
-# now determine Cube Spaxel flux
-            IFUCube = CubeData.setup_IFUCube(self,j)
-
-            CubeData.update_IFUCube(self,IFUCube, spaxel)
-#            print('***** ',IFUCube.meta.wcsinfo.crval1,IFUCube.meta.wcsinfo.crval2,
-#                  IFUCube.meta.wcsinfo.crval3)
-
-#            print('ra,dec,lambda 1,1,1 ',IFUCube.meta.wcs(1,1,1))
-            
-
-            t1 = time.time()
-            log.info("Time Create Single IFUcube  = %.1f.s" % (t1 - t0,))
-
-#_______________________________________________________________________
-            single_IFUCube.append(IFUCube)
-            del spaxel[:]
-        return single_IFUCube
-
-#********************************************************************************
-    def find_output_type(self):
-        
-        ValidChannel = ['1', '2', '3', '4']
-        ValidSubchannel = ['SHORT', 'MEDIUM', 'LONG']
-
-        nchannels = len(ValidChannel)
-        nsubchannels = len(ValidSubchannel)
-
-
-        ValidGWA = ['G140M', 'G140H', 'G140M', 'G140H', 'G235M', 'G235H',
-                    'G395M', 'G395H', 'PRISM']
-        ValidFWA = ['F070LP', 'F070LP', 'F100LP', 'F100LP', 'F170LP',
-                    'F170LP', 'F290LP', 'F290LP', 'CLEAR']
-
-        nbands = len(ValidFWA)
-
-
-
-#********************************************************************************
-    def create_spaxel(self):
-        """
-        Short Summary
-        -------------
-        # now you have the size of cube - create an instance for each spaxel
-        # create an empty spaxel list - this will become a list of Spaxel classses
-
-        Parameter
-        ----------
-
-        Returns
-        -------
-        list of classes contained in spaxel
-        """
-#________________________________________________________________________________
-
-
-        total_num = self.naxis1*self.naxis2*self.naxis3
-
-        if(self.interpolation == 'pointcloud'):
-            for t in range(total_num):
-                self.spaxel.append(spaxel.Spaxel())
-        else:
-            for t in range(total_num):
-                self.spaxel.append(spaxel.SpaxelAB())
-
-        return self.spaxel
-
-#********************************************************************************
-    def determine_roi_size(self):
-        """
-        Short Summary
-        -------------
-        Determine the spatial and wavelength roi size to use for selecting point cloud
-        elements around the spaxel centeres.
-        If the IFU cube covers more than 1 band - then use the rules to
-        define the Spatial and Wavelength roi size to use for the cube
-        Current Rule: using the minimum
-
-        Parameters
-        ----------
-        instrument_info holds the defaults roi sizes  for each channel/subchannel (MIRI)
-        or grating (NIRSPEC)
-
-        Returns
-        -------
-        roi size for spatial and wavelength
-
-        """
-
-        roi = [0, 0]
-        if self.instrument == 'MIRI':
-            number_bands = len(self.band_channel)
-            min_s = 1000.00
-            min_w = 1000.00
-
-            for i in range(number_bands):
-                this_channel = self.band_channel[i]
-                this_sub = self.band_subchannel[i]
-                wroi = self.instrument_info.GetWaveRoi(this_channel,this_sub)
-                if wroi < min_w:
-                    min_w = wroi
-                sroi = self.instrument_info.GetSpatialRoi(this_channel,this_sub)
-                if sroi < min_s:
-                    min_s = sroi
-            roi = [min_w, min_s]
-
-        elif self.instrument == 'NIRSPEC':
-            number_gratings = len(self.band_grating)
-            min_s = 1000.00
-            min_w = 1000.00
-
-            for i in range(number_gratings):
-                this_gwa = self.band_grating[i]
-                wroi = self.instrument_info.GetWaveRoi(this_gwa)
-                if wroi < min_w:
-                    min_w = wroi
-                sroi = self.instrument_info.GetSpatialRoi(this_gwa)
-                if sroi < min_s:
-                    min_s = sroi
-            roi = [min_w, min_s]
-        return roi
-
-#********************************************************************************
-
-#    def subtract_background(self)
-
-
-#********************************************************************************
-
-    def map_detector_to_spaxel(self,this_par1, this_par2,spaxel):
-        from ..mrs_imatch.mrs_imatch_step import apply_background_2d 
-#********************************************************************************
-        """
-        Short Summary
-        -------------
-        Loop over files that cover the cube and map the detector pixel to Cube spaxels
-        If dither offsets have been supplied then apply those values to the data
-
-        Parameter
-        ----------
-        spaxel: List of Spaxels
-
-        Returns
-        -------
-        if(interpolation = area - only valid for alpha-beta
-        or
-        if(interpolation = pointcloud
-        """
-
-        instrument  = self.instrument
-        nfiles = len(self.master_table.FileMap[instrument][this_par1][this_par2])
-        log.info('Number of files/models in band %i', nfiles)
-
-    # loop over the files that cover the spectral range the cube is for
-
-        for k in range(nfiles):
-            ifile = self.master_table.FileMap[instrument][this_par1][this_par2][k]
-            ioffset = len(self.master_table.FileOffset[this_par1][this_par2]['C1'])
-
-            c1_offset = 0.0
-            c2_offset = 0.0
-        # c1_offset and c2_offset are the dither offset sets (in arc seconds)
-        # by default these are zer0. The user has to supply these
-            if ioffset == nfiles:
-                c1_offset = self.master_table.FileOffset[this_par1][this_par2]['C1'][k]
-                c2_offset = self.master_table.FileOffset[this_par1][this_par2]['C2'][k]
-# Open the input data model
-            with datamodels.ImageModel(ifile) as input_model:
-                # check if background sky matching as been done
-                # mrs_imatch step. THis is only for MRS data at this time
-                # but go head and check it before splitting by instrument
-                # the polynomial should be empty for NIRSPEC
-                do_background_subtraction = False
-                num_ch_bgk = len(input_model.meta.background.polynomial_info)
-
-                if(num_ch_bgk> 0):
-
-                    do_background_subtraction = True
-                    for ich_num in range(num_ch_bgk):
-                        poly = input_model.meta.background.polynomial_info[ich_num]
-                        poly_ch = poly.channel
-                        if(poly_ch == this_par1):
-                            print('*****Going to subtract background *******')
-                            print('Channel',this_par1)
-                            apply_background_2d(input_model,poly_ch,subtract=True)
-
-                
-#********************************************************************************
-                if self.instrument == 'MIRI':
-#________________________________________________________________________________
-# Standard method
-                    if self.interpolation == 'pointcloud':
-                        xstart, xend = self.instrument_info.GetMIRISliceEndPts(this_par1)
-                        y, x = np.mgrid[:1024, xstart:xend]
-                        y = np.reshape(y, y.size)
-                        x = np.reshape(x, x.size)
-                        t0 = time.time()
-                        cube_cloud.match_det2cube(self,input_model,
-                                            x, y, k,
-                                            this_par1,this_par2,
-                                            spaxel,
-                                            c1_offset, c2_offset)
-
-
-                        t1 = time.time()
-                        log.debug("Time Match one Channel from 1 file  to IFUCube = %.1f.s"
-                                  % (t1 - t0,))
-#________________________________________________________________________________
-#2D area method - only works for single files and coord_system = 'alpha-beta'
-                    if self.interpolation == 'area':
-                        det2ab_transform = input_model.meta.wcs.get_transform('detector',
-                                                                              'alpha_beta')
-
-
-                        start_region = self.instrument_info.GetStartSlice(this_par1)
-                        end_region = self.instrument_info.GetEndSlice(this_par1)
-                        regions = list(range(start_region, end_region + 1))
-
-                    #xtest = 28.310396-1 # test pixel to compare with Distortion doc
-                    #ytest = 512.0-1     # test pixel to compare with Distortion doc
-                    #coord1_test,coord2_test,lam_test = det2ab_shift(xtest,ytest)
-                    #print('test values',xtest+1,ytest+1,coord1_test,coord2_test,lam_test)
-
-                        for i in regions:
-                            log.info('Working on Slice # %d', i)
-
-                            y, x = (det2ab_transform.label_mapper.mapper == i).nonzero()
-
-                    # spaxel object holds all needed information in a set of lists
-                    #    flux (of overlapping detector pixel)
-                    #    error (of overlapping detector pixel)
-                    #    overlap ratio
-                    #    beta distance
-
-# getting pixel corner - ytop = y + 1 (routine fails for y = 1024)
-                            index = np.where(y < 1023)
-                            y = y[index]
-                            x = x[index]
-                            t0 = time.time()
-
-
-                            cube_overlap.match_det2cube(self, x, y, i,
-                                                        start_region,
-                                                        input_model,
-                                                        det2ab_transform,
-                                                        spaxel)
-                            t1 = time.time()
-                            log.debug("Time Map one Slice  to Cube = %.1f.s" % (t1 - t0,))
-
-#********************************************************************************
-                elif instrument == 'NIRSPEC':
-                    # each file, detector has 30 slices - wcs information access seperately for each slice
-                    start_slice = 0
-                    end_slice = 29
-                    nslices = end_slice - start_slice + 1
-                    regions = list(range(start_slice, end_slice + 1))
-                    log.info("Mapping each NIRSPEC slice to sky, this takes a while for NIRSPEC data")
-                    for i in regions:
-#                    print('on region ',i)
-                        slice_wcs = nirspec.nrs_wcs_set_input(input_model, i)
-                        yrange = slice_wcs.bounding_box[1][0],slice_wcs.bounding_box[1][1]
-                        xrange = slice_wcs.bounding_box[0][0],slice_wcs.bounding_box[0][1]
-
-
-                        x,y = wcstools.grid_from_bounding_box(slice_wcs.bounding_box,
-                                                              step=(1,1), center=True)
-                        t0 = time.time()
-                        cube_cloud.match_det2cube(self,input_model,
-                                                  x, y, i,
-                                                  this_par1,this_par2,
-                                                  spaxel,
-                                                  c1_offset, c2_offset)
-
-
-                        t1 = time.time()
-                        log.debug("Time Match one NIRSPEC slice  to IFUCube = %.1f.s" % (t1 - t0,))
-#********************************************************************************
-    def find_spaxel_flux(self, spaxel):
-#********************************************************************************
-        """
-        Short Summary
-        -------------
-        Depending on the interpolation method, find the flux for each spaxel value
-
-        Parameter
-        ----------
-        spaxel: List of Spaxels
-        PixelCloud - pixel point cloud, only filled in if doing 3-D interpolation
-
-        Returns
-        -------
-        if(interpolation = area) flux determined for each spaxel
-        or
-        if(interpolation = pointcloud) flux determined for each spaxel based on interpolation of PixelCloud
-        """
-
-
-        if self.interpolation == 'area':
-            nspaxel = len(spaxel)
-
-            for i in range(nspaxel):
-                if(spaxel[i].iflux > 0):
-                    spaxel[i].flux = spaxel[i].flux/spaxel[i].flux_weight
-
-        elif self.interpolation == 'pointcloud':
-            icube = 0
-            t0 = time.time()
-            for iz, z in enumerate(self.zcoord):
-                for iy, y in enumerate(self.ycoord):
-                    for ix, x in enumerate(self.xcoord):
-
-                        if(spaxel[icube].iflux > 0):
-                            spaxel[icube].flux = spaxel[icube].flux/spaxel[icube].flux_weight
-
-                            if(self.debug_pixel == 1 and self.xdebug == ix and
-                               self.ydebug == iy and self.zdebug == iz ):
-
-                                log.debug('For spaxel %d %d %d final flux %f '
-                                          %(self.xdebug+1,self.ydebug+1,
-                                            self.zdebug+1,spaxel[icube].flux))
-                                self.spaxel_debug.write('For spaxel %d %d %d, final flux %f '
-                                                        %(self.xdebug+1,self.ydebug+1,
-                                                          self.zdebug+1,spaxel[icube].flux) +' \n')
-                        icube = icube + 1
-            t1 = time.time()
-            log.info("Time to interpolate at spaxel values = %.1f.s" % (t1 - t0,))
-
-#********************************************************************************
-    def setup_IFUCube(self,j):
-
-        """
-        Short Summary
-        -------------
-        Set up the final  the IFU cube to fits file
-
-        Parameters
-        ----------
-        Cube: holds meta data of cube
-        spaxel: list of spaxels in cube
-
-
-        Returns
-        -------
-        return IFUCube model
-
-        """
-        naxis1 = self.naxis1
-        naxis2 = self.naxis2
-        naxis3 = self.naxis3
-
-
-        data = np.zeros((naxis3, naxis2, naxis1))
-        idata = np.zeros((naxis3, naxis2, naxis1))
-
-        dq_cube = np.zeros((naxis3, naxis2, naxis1))
-        err_cube = np.zeros((naxis3, naxis2, naxis1))
-
-        IFUCube = datamodels.IFUCubeModel(data=data, dq=dq_cube, err=err_cube, weightmap=idata)
-        IFUCube.update(self.input_models[j])
-
-        IFUCube.meta.filename = self.output_name
-        if self.single:
-            with datamodels.open(self.input_models[j]) as input:
-                # makingf fileanme = org gives a error later when past
-                # back to model container - do we want to define
-                # a new KEYWORD - filename_org ?
-                #IFUCube.meta.filename = input.meta.filename
-
-                filename = self.input_filenames[j]
-                indx = filename.rfind('.fits')
-                self.output_name_base = filename[:indx]
-                self.output_file = None
-                newname  = cube_build_io_util.update_output_name(self)
-                IFUCube.meta.filename = newname
-                IFUCube.meta.instrument.channel = self.band_channel[0] 
-
-        IFUCube.meta.wcsinfo.crval1 = self.Crval1
-        IFUCube.meta.wcsinfo.crval2 = self.Crval2
-        IFUCube.meta.wcsinfo.crval3 = self.Crval3
-        IFUCube.meta.wcsinfo.crpix1 = self.Crpix1
-        IFUCube.meta.wcsinfo.crpix2 = self.Crpix2
-        IFUCube.meta.wcsinfo.crpix3 = self.Crpix3
-        IFUCube.meta.wcsinfo.cdelt1 = self.Cdelt1/3600.0
-        IFUCube.meta.wcsinfo.cdelt2 = self.Cdelt2/3600.0
-        IFUCube.meta.wcsinfo.cdelt3 = self.Cdelt3
-
-        IFUCube.meta.wcsinfo.ctype1 = 'RA---TAN'
-        IFUCube.meta.wcsinfo.ctype2 = 'DEC--TAN'
-        IFUCube.meta.wcsinfo.cunit1 = 'deg'
-        IFUCube.meta.wcsinfo.cunit2 = 'deg'
-
-        IFUCube.meta.wcsinfo.ctype3 = 'WAVE'
-        IFUCube.meta.wcsinfo.cunit3 = 'um'
-        IFUCube.meta.wcsinfo.wcsaxes = 3
-        IFUCube.meta.wcsinfo.pc1_1 = 1
-        IFUCube.meta.wcsinfo.pc1_2 = 0
-        IFUCube.meta.wcsinfo.pc1_3 = 0
-
-        IFUCube.meta.wcsinfo.pc2_1 = 0
-        IFUCube.meta.wcsinfo.pc2_2 = 1
-        IFUCube.meta.wcsinfo.pc2_3 = 0
-
-        IFUCube.meta.wcsinfo.pc3_1 = 0
-        IFUCube.meta.wcsinfo.pc3_2 = 0
-        IFUCube.meta.wcsinfo.pc3_3 = 1
-
-        IFUCube.meta.flux_extension = 'SCI'
-        IFUCube.meta.error_extension = 'ERR'
-        IFUCube.meta.dq_extension = 'DQ'
-        IFUCube.meta.roi_spatial = self.rois    
-        IFUCube.meta.roi_wave = self.roiw
-        IFUCube.meta.weighting = self.weighting
-        IFUCube.meta.weight_power = self.weight_power
-
-#        IFUCube.meta.data_model_type = 'IFUCubeModel'
-        IFUCube.error_type = 'ERR'
-
-        if self.coord_system == 'alpha-beta' :
-            IFUCube.meta.wcsinfo.cunit1 = 'arcsec'
-            IFUCube.meta.wcsinfo.cunit2 = 'arcsec'
-
-            if self.channel == '1' and self.band == 'SHORT' :
-                IFUCube.meta.wcsinfo.ctype1 = 'MRSAL1A'
-                IFUCube.meta.wcsinfo.ctype2 = 'MRSBE1A'
-            if self.channel == '2' and self.band == 'SHORT':
-                IFUCube.meta.wcsinfo.ctype1 = 'MRSAL2A'
-                IFUCube.meta.wcsinfo.ctype2 = 'MRSBE2A'
-            if self.channel == '3' and self.band == 'SHORT':
-                IFUCube.meta.wcsinfo.ctype1 = 'MRSAL3A'
-                IFUCube.meta.wcsinfo.ctype2 = 'MRSBE3A'
-            if self.channel == '4' and self.band == 'SHORT':
-                IFUCube.meta.wcsinfo.ctype1 = 'MRSAL4A'
-                IFUCube.meta.wcsinfo.ctype2 = 'MRSBE4A'
-
-            if self.channel == '1' and self.band == 'MEDIUM':
-                IFUCube.meta.wcsinfo.ctype1 = 'MRSAL1B'
-                IFUCube.meta.wcsinfo.ctype2 = 'MRSBE1B'
-            if self.channel == '2' and self.band == 'MEDIUM':
-                IFUCube.meta.wcsinfo.ctype1 = 'MRSAL2B'
-                IFUCube.meta.wcsinfo.ctype2 = 'MRSBE2B'
-            if self.channel == '3' and self.band == 'MEDIUM':
-                IFUCube.meta.wcsinfo.ctype1 = 'MRSAL3B'
-                IFUCube.meta.wcsinfo.ctype2 = 'MRSBE3B'
-            if self.channel == '4' and self.band == 'MEDIUM':
-                IFUCube.meta.wcsinfo.ctype1 = 'MRSAL4B'
-                IFUCube.meta.wcsinfo.ctype2 = 'MRSBE4B'
-
-            if self.channel == '1' and self.band == 'LONG':
-                IFUCube.meta.wcsinfo.ctype1 = 'MRSAL1C'
-                IFUCube.meta.wcsinfo.ctype2 = 'MRSBE1C'
-            if self.channel == '2' and self.band == 'LONG':
-                IFUCube.meta.wcsinfo.ctype1 = 'MRSAL2C'
-                IFUCube.meta.wcsinfo.ctype2 = 'MRSBE2C'
-            if self.channel == '3' and self.band == 'LONG':
-                IFUCube.meta.wcsinfo.ctype1 = 'MRSAL3C'
-                IFUCube.meta.wcsinfo.ctype2 = 'MRSBE3C'
-            if self.channel == '4' and self.band == 'LONG':
-                IFUCube.meta.wcsinfo.ctype1 = 'MRSAL4C'
-                IFUCube.meta.wcsinfo.ctype2 = 'MRSBE4C'
-
-
-
-
-        wcsobj = pointing.create_fitswcs(IFUCube)
-        IFUCube.meta.wcs = wcsobj
-        IFUCube.meta.wcs.bounding_box = ((0,naxis1-1),(0,naxis2-1),(0,naxis3-1))
-        return IFUCube
-
-#********************************************************************************
-
-    def update_IFUCube(self,IFUCube, spaxel):
-
-#********************************************************************************
-        """
-        Short Summary
-        -------------
-        Write the IFU cube to fits file
-
-        Parameters
-        ----------
-        Cube: holds meta data of cube
-        spaxel: list of spaxels in cube
-
-
-        Returns
-        -------
-        fills in IFUdata arrays with spaxel
-
-        """
-    #pull out data into array
-
-
-        temp_flux =np.reshape(np.array([s.flux for s in spaxel]),
-                          [self.naxis3,self.naxis2,self.naxis1])
-        temp_wmap =np.reshape(np.array([s.iflux for s in spaxel]),
-                          [self.naxis3,self.naxis2,self.naxis1])
-
-
-        IFUCube.data = temp_flux
-        IFUCube.weightmap = temp_wmap
-
-        IFUCube.meta.cal_step.cube_build = 'COMPLETE'
-#    icube = 0
-#    for z in range(Cube.naxis3):
-#        for y in range(Cube.naxis2):
-#            for x in range(Cube.naxis1):
-#                IFUCube.data[z, y, x] = spaxel[icube].flux
-#                IFUCube.weightmap[z, y, x] = len(spaxel[icube].ipointcloud)
-#                icube = icube + 1
-
-
-       # result = IFUCube.copy()
-        #return result
-
-#********************************************************************************
